requirements.txt
pyyaml>=6.0
python-docx>=1.0
reportlab>=4.0
lxml>=4.9
